+++ lib/constructorio.js
@@ -1,30 +1,37 @@
+"use strict";
+
+function _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError("Cannot call a class as a function"); } }
+
+function _defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if ("value" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } }
+
+function _createClass(Constructor, protoProps, staticProps) { if (protoProps) _defineProperties(Constructor.prototype, protoProps); if (staticProps) _defineProperties(Constructor, staticProps); return Constructor; }
+
 /* eslint-disable camelcase, no-unneeded-ternary */
-const ConstructorioID = require('@constructor-io/constructorio-id'); // Modules
+var ConstructorioID = require('@constructor-io/constructorio-id'); // Modules
 
 
-const Search = require('./modules/search');
+var Search = require('./modules/search');
 
-const Browse = require('./modules/browse');
+var Browse = require('./modules/browse');
 
-const Autocomplete = require('./modules/autocomplete');
+var Autocomplete = require('./modules/autocomplete');
 
-const Recommendations = require('./modules/recommendations');
+var Recommendations = require('./modules/recommendations');
 
-const Tracker = require('./modules/tracker');
+var Tracker = require('./modules/tracker');
 
-const EventDispatcher = require('./utils/event-dispatcher');
+var EventDispatcher = require('./utils/event-dispatcher');
 
-const helpers = require('./utils/helpers');
+var helpers = require('./utils/helpers');
 
-const {
-  version: packageVersion
-} = require('../package.json');
+var _require = require('../package.json'),
+    packageVersion = _require.version;
 /**
  * Class to instantiate the ConstructorIO client.
  */
 
 
-class ConstructorIO {
+var ConstructorIO = /*#__PURE__*/function () {
   /**
    * @param {string} apiKey - Constructor.io API key
    * @param {string} [serviceUrl='https://ac.cnstrc.com'] - API URL endpoint
@@ -48,53 +55,55 @@ class ConstructorIO {
    * @property {object} [tracker] - Interface to {@link module:tracker}
    * @returns {class}
    */
-  constructor(options = {}) {
-    const {
-      apiKey,
-      version,
-      serviceUrl,
-      segments,
-      testCells,
-      clientId,
-      sessionId,
-      userId,
-      fetch,
-      trackingSendDelay,
-      sendReferrerWithTrackingEvents,
-      sendTrackingEvents,
-      eventDispatcher,
-      idOptions,
-      beaconMode
-    } = options;
+  function ConstructorIO() {
+    var options = arguments.length > 0 && arguments[0] !== undefined ? arguments[0] : {};
+
+    _classCallCheck(this, ConstructorIO);
+
+    var apiKey = options.apiKey,
+        version = options.version,
+        serviceUrl = options.serviceUrl,
+        segments = options.segments,
+        testCells = options.testCells,
+        clientId = options.clientId,
+        sessionId = options.sessionId,
+        userId = options.userId,
+        fetch = options.fetch,
+        trackingSendDelay = options.trackingSendDelay,
+        sendReferrerWithTrackingEvents = options.sendReferrerWithTrackingEvents,
+        sendTrackingEvents = options.sendTrackingEvents,
+        eventDispatcher = options.eventDispatcher,
+        idOptions = options.idOptions,
+        beaconMode = options.beaconMode;
 
     if (!apiKey || typeof apiKey !== 'string') {
       throw new Error('API key is a required parameter of type string');
     }
 
-    let session_id;
-    let client_id; // Initialize ID session (if within browser context)
+    var session_id;
+    var client_id; // Initialize ID session (if within browser context)
 
     if (helpers.hasWindow()) {
-      ({
-        session_id,
-        client_id
-      } = new ConstructorioID(idOptions || {}));
+      var _ConstructorioID = new ConstructorioID(idOptions || {});
+
+      session_id = _ConstructorioID.session_id;
+      client_id = _ConstructorioID.client_id;
     }
 
     this.options = {
-      apiKey,
-      version: version || global.CLIENT_VERSION || `ciojs-client-${packageVersion}`,
+      apiKey: apiKey,
+      version: version || global.CLIENT_VERSION || "ciojs-client-".concat(packageVersion),
       serviceUrl: serviceUrl || 'https://ac.cnstrc.com',
       sessionId: sessionId || session_id,
       clientId: clientId || client_id,
-      userId,
-      segments,
-      testCells,
-      fetch,
-      trackingSendDelay,
-      sendTrackingEvents,
-      sendReferrerWithTrackingEvents,
-      eventDispatcher,
+      userId: userId,
+      segments: segments,
+      testCells: testCells,
+      fetch: fetch,
+      trackingSendDelay: trackingSendDelay,
+      sendTrackingEvents: sendTrackingEvents,
+      sendReferrerWithTrackingEvents: sendReferrerWithTrackingEvents,
+      eventDispatcher: eventDispatcher,
       beaconMode: beaconMode === false ? false : true // Defaults to 'true',
 
     }; // Disable event dispatcher and tracking events if `window` not available
@@ -128,33 +137,35 @@ class ConstructorIO {
    */
 
 
-  setClientOptions(options) {
-    if (Object.keys(options).length) {
-      const {
-        apiKey,
-        segments,
-        testCells,
-        userId
-      } = options;
+  _createClass(ConstructorIO, [{
+    key: "setClientOptions",
+    value: function setClientOptions(options) {
+      if (Object.keys(options).length) {
+        var apiKey = options.apiKey,
+            segments = options.segments,
+            testCells = options.testCells,
+            userId = options.userId;
 
-      if (apiKey) {
-        this.options.apiKey = apiKey;
-      }
+        if (apiKey) {
+          this.options.apiKey = apiKey;
+        }
 
-      if (segments) {
-        this.options.segments = segments;
-      }
+        if (segments) {
+          this.options.segments = segments;
+        }
 
-      if (testCells) {
-        this.options.testCells = testCells;
-      }
+        if (testCells) {
+          this.options.testCells = testCells;
+        }
 
-      if (userId) {
-        this.options.userId = userId;
+        if (userId) {
+          this.options.userId = userId;
+        }
       }
     }
-  }
+  }]);
 
-}
+  return ConstructorIO;
+}();
 
 module.exports = ConstructorIO;
