+++ lib/modules/autocomplete.js
@@ -1,30 +1,35 @@
+"use strict";
+
+function _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError("Cannot call a class as a function"); } }
+
+function _defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if ("value" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } }
+
+function _createClass(Constructor, protoProps, staticProps) { if (protoProps) _defineProperties(Constructor.prototype, protoProps); if (staticProps) _defineProperties(Constructor, staticProps); return Constructor; }
+
 /* eslint-disable object-curly-newline, no-underscore-dangle */
-const qs = require('qs');
+var qs = require('qs');
 
-const fetchPonyfill = require('fetch-ponyfill');
+var fetchPonyfill = require('fetch-ponyfill');
 
-const Promise = require('es6-promise');
+var Promise = require('es6-promise');
 
-const EventDispatcher = require('../utils/event-dispatcher');
+var EventDispatcher = require('../utils/event-dispatcher');
 
-const {
-  throwHttpErrorFromResponse,
-  cleanParams
-} = require('../utils/helpers'); // Create URL from supplied query (term) and parameters
+var _require = require('../utils/helpers'),
+    throwHttpErrorFromResponse = _require.throwHttpErrorFromResponse,
+    cleanParams = _require.cleanParams; // Create URL from supplied query (term) and parameters
 
 
 function createAutocompleteUrl(query, parameters, options) {
-  const {
-    apiKey,
-    version,
-    serviceUrl,
-    sessionId,
-    clientId,
-    userId,
-    segments,
-    testCells
-  } = options;
-  let queryParams = {
+  var apiKey = options.apiKey,
+      version = options.version,
+      serviceUrl = options.serviceUrl,
+      sessionId = options.sessionId,
+      clientId = options.clientId,
+      userId = options.userId,
+      segments = options.segments,
+      testCells = options.testCells;
+  var queryParams = {
     c: version
   };
   queryParams.key = apiKey;
@@ -37,8 +42,8 @@ function createAutocompleteUrl(query, parameters, options) {
 
 
   if (testCells) {
-    Object.keys(testCells).forEach(testCellKey => {
-      queryParams[`ef-${testCellKey}`] = testCells[testCellKey];
+    Object.keys(testCells).forEach(function (testCellKey) {
+      queryParams["ef-".concat(testCellKey)] = testCells[testCellKey];
     });
   } // Pull user segments from options
 
@@ -53,11 +58,9 @@ function createAutocompleteUrl(query, parameters, options) {
   }
 
   if (parameters) {
-    const {
-      numResults,
-      resultsPerSection,
-      filters
-    } = parameters; // Pull results number from parameters
+    var numResults = parameters.numResults,
+        resultsPerSection = parameters.resultsPerSection,
+        filters = parameters.filters; // Pull results number from parameters
 
     if (numResults) {
       queryParams.num_results = numResults;
@@ -65,8 +68,8 @@ function createAutocompleteUrl(query, parameters, options) {
 
 
     if (resultsPerSection) {
-      Object.keys(resultsPerSection).forEach(section => {
-        queryParams[`num_results_${section}`] = resultsPerSection[section];
+      Object.keys(resultsPerSection).forEach(function (section) {
+        queryParams["num_results_".concat(section)] = resultsPerSection[section];
       });
     } // Pull filters from parameters
 
@@ -78,10 +81,10 @@ function createAutocompleteUrl(query, parameters, options) {
 
   queryParams._dt = Date.now();
   queryParams = cleanParams(queryParams);
-  const queryString = qs.stringify(queryParams, {
+  var queryString = qs.stringify(queryParams, {
     indices: false
   });
-  return `${serviceUrl}/autocomplete/${encodeURIComponent(query)}?${queryString}`;
+  return "".concat(serviceUrl, "/autocomplete/").concat(encodeURIComponent(query), "?").concat(queryString);
 }
 /**
  * Interface to autocomplete related API calls.
@@ -92,8 +95,10 @@ function createAutocompleteUrl(query, parameters, options) {
  */
 
 
-class Autocomplete {
-  constructor(options) {
+var Autocomplete = /*#__PURE__*/function () {
+  function Autocomplete(options) {
+    _classCallCheck(this, Autocomplete);
+
     this.options = options || {};
     this.eventDispatcher = new EventDispatcher(options.eventDispatcher);
   }
@@ -110,49 +115,56 @@ class Autocomplete {
    */
 
 
-  getAutocompleteResults(query, parameters) {
-    let requestUrl;
-    const fetch = this.options && this.options.fetch || fetchPonyfill({
-      Promise
-    }).fetch;
+  _createClass(Autocomplete, [{
+    key: "getAutocompleteResults",
+    value: function getAutocompleteResults(query, parameters) {
+      var _this = this;
 
-    try {
-      requestUrl = createAutocompleteUrl(query, parameters, this.options);
-    } catch (e) {
-      return Promise.reject(e);
-    }
+      var requestUrl;
+      var fetch = this.options && this.options.fetch || fetchPonyfill({
+        Promise: Promise
+      }).fetch;
 
-    return fetch(requestUrl).then(response => {
-      if (response.ok) {
-        return response.json();
+      try {
+        requestUrl = createAutocompleteUrl(query, parameters, this.options);
+      } catch (e) {
+        return Promise.reject(e);
       }
 
-      return throwHttpErrorFromResponse(new Error(), response);
-    }).then(json => {
-      if (json.sections) {
-        if (json.result_id) {
-          const sectionKeys = Object.keys(json.sections);
-          sectionKeys.forEach(section => {
-            const sectionItems = json.sections[section];
-
-            if (sectionItems.length) {
-              // Append `result_id` to each section item
-              sectionItems.forEach(item => {
-                // eslint-disable-next-line no-param-reassign
-                item.result_id = json.result_id;
-              });
-            }
-          });
+      return fetch(requestUrl).then(function (response) {
+        if (response.ok) {
+          return response.json();
         }
 
-        this.eventDispatcher.queue('autocomplete.getAutocompleteResults.completed', json);
-        return json;
-      }
+        return throwHttpErrorFromResponse(new Error(), response);
+      }).then(function (json) {
+        if (json.sections) {
+          if (json.result_id) {
+            var sectionKeys = Object.keys(json.sections);
+            sectionKeys.forEach(function (section) {
+              var sectionItems = json.sections[section];
 
-      throw new Error('getAutocompleteResults response data is malformed');
-    });
-  }
+              if (sectionItems.length) {
+                // Append `result_id` to each section item
+                sectionItems.forEach(function (item) {
+                  // eslint-disable-next-line no-param-reassign
+                  item.result_id = json.result_id;
+                });
+              }
+            });
+          }
 
-}
+          _this.eventDispatcher.queue('autocomplete.getAutocompleteResults.completed', json);
+
+          return json;
+        }
+
+        throw new Error('getAutocompleteResults response data is malformed');
+      });
+    }
+  }]);
+
+  return Autocomplete;
+}();
 
 module.exports = Autocomplete;
