+++ lib/modules/browse.js
@@ -1,27 +1,33 @@
+"use strict";
+
+function _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError("Cannot call a class as a function"); } }
+
+function _defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if ("value" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } }
+
+function _createClass(Constructor, protoProps, staticProps) { if (protoProps) _defineProperties(Constructor.prototype, protoProps); if (staticProps) _defineProperties(Constructor, staticProps); return Constructor; }
+
 /* eslint-disable object-curly-newline, no-underscore-dangle */
-const qs = require('qs');
+var qs = require('qs');
 
-const fetchPonyfill = require('fetch-ponyfill');
+var fetchPonyfill = require('fetch-ponyfill');
 
-const Promise = require('es6-promise');
+var Promise = require('es6-promise');
 
-const EventDispatcher = require('../utils/event-dispatcher');
+var EventDispatcher = require('../utils/event-dispatcher');
 
-const helpers = require('../utils/helpers'); // Create URL from supplied filter name, value and parameters
+var helpers = require('../utils/helpers'); // Create URL from supplied filter name, value and parameters
 
 
 function createBrowseUrl(filterName, filterValue, parameters, options) {
-  const {
-    apiKey,
-    version,
-    serviceUrl,
-    sessionId,
-    clientId,
-    userId,
-    segments,
-    testCells
-  } = options;
-  let queryParams = {
+  var apiKey = options.apiKey,
+      version = options.version,
+      serviceUrl = options.serviceUrl,
+      sessionId = options.sessionId,
+      clientId = options.clientId,
+      userId = options.userId,
+      segments = options.segments,
+      testCells = options.testCells;
+  var queryParams = {
     c: version
   };
   queryParams.key = apiKey;
@@ -39,8 +45,8 @@ function createBrowseUrl(filterName, filterValue, parameters, options) {
 
 
   if (testCells) {
-    Object.keys(testCells).forEach(testCellKey => {
-      queryParams[`ef-${testCellKey}`] = testCells[testCellKey];
+    Object.keys(testCells).forEach(function (testCellKey) {
+      queryParams["ef-".concat(testCellKey)] = testCells[testCellKey];
     });
   } // Pull user segments from options
 
@@ -55,15 +61,13 @@ function createBrowseUrl(filterName, filterValue, parameters, options) {
   }
 
   if (parameters) {
-    const {
-      page,
-      resultsPerPage,
-      filters,
-      sortBy,
-      sortOrder,
-      section,
-      fmtOptions
-    } = parameters; // Pull page from parameters
+    var page = parameters.page,
+        resultsPerPage = parameters.resultsPerPage,
+        filters = parameters.filters,
+        sortBy = parameters.sortBy,
+        sortOrder = parameters.sortOrder,
+        section = parameters.section,
+        fmtOptions = parameters.fmtOptions; // Pull page from parameters
 
     if (!helpers.isNil(page)) {
       queryParams.page = page;
@@ -101,10 +105,10 @@ function createBrowseUrl(filterName, filterValue, parameters, options) {
 
   queryParams._dt = Date.now();
   queryParams = helpers.cleanParams(queryParams);
-  const queryString = qs.stringify(queryParams, {
+  var queryString = qs.stringify(queryParams, {
     indices: false
   });
-  return `${serviceUrl}/browse/${encodeURIComponent(filterName)}/${encodeURIComponent(filterValue)}?${queryString}`;
+  return "".concat(serviceUrl, "/browse/").concat(encodeURIComponent(filterName), "/").concat(encodeURIComponent(filterValue), "?").concat(queryString);
 }
 /**
  * Interface to browse related API calls
@@ -115,8 +119,10 @@ function createBrowseUrl(filterName, filterValue, parameters, options) {
  */
 
 
-class Browse {
-  constructor(options) {
+var Browse = /*#__PURE__*/function () {
+  function Browse(options) {
+    _classCallCheck(this, Browse);
+
     this.options = options || {};
     this.eventDispatcher = new EventDispatcher(options.eventDispatcher);
   }
@@ -138,42 +144,49 @@ class Browse {
    */
 
 
-  getBrowseResults(filterName, filterValue, parameters) {
-    let requestUrl;
-    const fetch = this.options && this.options.fetch || fetchPonyfill({
-      Promise
-    }).fetch;
+  _createClass(Browse, [{
+    key: "getBrowseResults",
+    value: function getBrowseResults(filterName, filterValue, parameters) {
+      var _this = this;
 
-    try {
-      requestUrl = createBrowseUrl(filterName, filterValue, parameters, this.options);
-    } catch (e) {
-      return Promise.reject(e);
-    }
+      var requestUrl;
+      var fetch = this.options && this.options.fetch || fetchPonyfill({
+        Promise: Promise
+      }).fetch;
 
-    return fetch(requestUrl).then(response => {
-      if (response.ok) {
-        return response.json();
+      try {
+        requestUrl = createBrowseUrl(filterName, filterValue, parameters, this.options);
+      } catch (e) {
+        return Promise.reject(e);
       }
 
-      return helpers.throwHttpErrorFromResponse(new Error(), response);
-    }).then(json => {
-      if (json.response && json.response.results) {
-        if (json.result_id) {
-          // Append `result_id` to each result item
-          json.response.results.forEach(result => {
-            // eslint-disable-next-line no-param-reassign
-            result.result_id = json.result_id;
-          });
+      return fetch(requestUrl).then(function (response) {
+        if (response.ok) {
+          return response.json();
         }
 
-        this.eventDispatcher.queue('browse.getBrowseResults.completed', json);
-        return json;
-      }
+        return helpers.throwHttpErrorFromResponse(new Error(), response);
+      }).then(function (json) {
+        if (json.response && json.response.results) {
+          if (json.result_id) {
+            // Append `result_id` to each result item
+            json.response.results.forEach(function (result) {
+              // eslint-disable-next-line no-param-reassign
+              result.result_id = json.result_id;
+            });
+          }
 
-      throw new Error('getBrowseResults response data is malformed');
-    });
-  }
+          _this.eventDispatcher.queue('browse.getBrowseResults.completed', json);
 
-}
+          return json;
+        }
+
+        throw new Error('getBrowseResults response data is malformed');
+      });
+    }
+  }]);
+
+  return Browse;
+}();
 
 module.exports = Browse;
