+++ lib/modules/recommendations.js
@@ -1,26 +1,32 @@
+"use strict";
+
+function _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError("Cannot call a class as a function"); } }
+
+function _defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if ("value" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } }
+
+function _createClass(Constructor, protoProps, staticProps) { if (protoProps) _defineProperties(Constructor.prototype, protoProps); if (staticProps) _defineProperties(Constructor, staticProps); return Constructor; }
+
 /* eslint-disable object-curly-newline, no-param-reassign */
-const qs = require('qs');
+var qs = require('qs');
 
-const fetchPonyfill = require('fetch-ponyfill');
+var fetchPonyfill = require('fetch-ponyfill');
 
-const Promise = require('es6-promise');
+var Promise = require('es6-promise');
 
-const EventDispatcher = require('../utils/event-dispatcher');
+var EventDispatcher = require('../utils/event-dispatcher');
 
-const helpers = require('../utils/helpers'); // Create URL from supplied parameters
+var helpers = require('../utils/helpers'); // Create URL from supplied parameters
 
 
 function createRecommendationsUrl(podId, parameters, options) {
-  const {
-    apiKey,
-    version,
-    serviceUrl,
-    sessionId,
-    userId,
-    clientId,
-    segments
-  } = options;
-  let queryParams = {
+  var apiKey = options.apiKey,
+      version = options.version,
+      serviceUrl = options.serviceUrl,
+      sessionId = options.sessionId,
+      userId = options.userId,
+      clientId = options.clientId,
+      segments = options.segments;
+  var queryParams = {
     c: version
   };
   queryParams.key = apiKey;
@@ -42,13 +48,11 @@ function createRecommendationsUrl(podId, parameters, options) {
   }
 
   if (parameters) {
-    const {
-      numResults,
-      itemIds,
-      section,
-      term,
-      filters
-    } = parameters; // Pull num results number from parameters
+    var numResults = parameters.numResults,
+        itemIds = parameters.itemIds,
+        section = parameters.section,
+        term = parameters.term,
+        filters = parameters.filters; // Pull num results number from parameters
 
     if (!helpers.isNil(numResults)) {
       queryParams.num_results = numResults;
@@ -76,10 +80,10 @@ function createRecommendationsUrl(podId, parameters, options) {
   }
 
   queryParams = helpers.cleanParams(queryParams);
-  const queryString = qs.stringify(queryParams, {
+  var queryString = qs.stringify(queryParams, {
     indices: false
   });
-  return `${serviceUrl}/recommendations/v1/pods/${podId}?${queryString}`;
+  return "".concat(serviceUrl, "/recommendations/v1/pods/").concat(podId, "?").concat(queryString);
 }
 /**
  * Interface to recommendations related API calls
@@ -90,8 +94,10 @@ function createRecommendationsUrl(podId, parameters, options) {
  */
 
 
-class Recommendations {
-  constructor(options) {
+var Recommendations = /*#__PURE__*/function () {
+  function Recommendations(options) {
+    _classCallCheck(this, Recommendations);
+
     this.options = options || {};
     this.eventDispatcher = new EventDispatcher(options.eventDispatcher);
   }
@@ -111,42 +117,49 @@ class Recommendations {
    */
 
 
-  getRecommendations(podId, parameters) {
-    let requestUrl;
-    const fetch = this.options && this.options.fetch || fetchPonyfill({
-      Promise
-    }).fetch;
-    parameters = parameters || {};
+  _createClass(Recommendations, [{
+    key: "getRecommendations",
+    value: function getRecommendations(podId, parameters) {
+      var _this = this;
 
-    try {
-      requestUrl = createRecommendationsUrl(podId, parameters, this.options);
-    } catch (e) {
-      return Promise.reject(e);
-    }
+      var requestUrl;
+      var fetch = this.options && this.options.fetch || fetchPonyfill({
+        Promise: Promise
+      }).fetch;
+      parameters = parameters || {};
 
-    return fetch(requestUrl).then(response => {
-      if (response.ok) {
-        return response.json();
+      try {
+        requestUrl = createRecommendationsUrl(podId, parameters, this.options);
+      } catch (e) {
+        return Promise.reject(e);
       }
 
-      return helpers.throwHttpErrorFromResponse(new Error(), response);
-    }).then(json => {
-      if (json.response && json.response.results) {
-        if (json.result_id) {
-          // Append `result_id` to each result item
-          json.response.results.forEach(result => {
-            result.result_id = json.result_id;
-          });
+      return fetch(requestUrl).then(function (response) {
+        if (response.ok) {
+          return response.json();
         }
 
-        this.eventDispatcher.queue('recommendations.getRecommendations.completed', json);
-        return json;
-      }
+        return helpers.throwHttpErrorFromResponse(new Error(), response);
+      }).then(function (json) {
+        if (json.response && json.response.results) {
+          if (json.result_id) {
+            // Append `result_id` to each result item
+            json.response.results.forEach(function (result) {
+              result.result_id = json.result_id;
+            });
+          }
 
-      throw new Error('getRecommendations response data is malformed');
-    });
-  }
+          _this.eventDispatcher.queue('recommendations.getRecommendations.completed', json);
 
-}
+          return json;
+        }
+
+        throw new Error('getRecommendations response data is malformed');
+      });
+    }
+  }]);
+
+  return Recommendations;
+}();
 
 module.exports = Recommendations;
