+++ lib/modules/search.js
@@ -1,27 +1,33 @@
+"use strict";
+
+function _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError("Cannot call a class as a function"); } }
+
+function _defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if ("value" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } }
+
+function _createClass(Constructor, protoProps, staticProps) { if (protoProps) _defineProperties(Constructor.prototype, protoProps); if (staticProps) _defineProperties(Constructor, staticProps); return Constructor; }
+
 /* eslint-disable object-curly-newline, no-underscore-dangle */
-const qs = require('qs');
+var qs = require('qs');
 
-const fetchPonyfill = require('fetch-ponyfill');
+var fetchPonyfill = require('fetch-ponyfill');
 
-const Promise = require('es6-promise');
+var Promise = require('es6-promise');
 
-const EventDispatcher = require('../utils/event-dispatcher');
+var EventDispatcher = require('../utils/event-dispatcher');
 
-const helpers = require('../utils/helpers'); // Create URL from supplied query (term) and parameters
+var helpers = require('../utils/helpers'); // Create URL from supplied query (term) and parameters
 
 
 function createSearchUrl(query, parameters, options) {
-  const {
-    apiKey,
-    version,
-    serviceUrl,
-    sessionId,
-    clientId,
-    userId,
-    segments,
-    testCells
-  } = options;
-  let queryParams = {
+  var apiKey = options.apiKey,
+      version = options.version,
+      serviceUrl = options.serviceUrl,
+      sessionId = options.sessionId,
+      clientId = options.clientId,
+      userId = options.userId,
+      segments = options.segments,
+      testCells = options.testCells;
+  var queryParams = {
     c: version
   };
   queryParams.key = apiKey;
@@ -34,8 +40,8 @@ function createSearchUrl(query, parameters, options) {
 
 
   if (testCells) {
-    Object.keys(testCells).forEach(testCellKey => {
-      queryParams[`ef-${testCellKey}`] = testCells[testCellKey];
+    Object.keys(testCells).forEach(function (testCellKey) {
+      queryParams["ef-".concat(testCellKey)] = testCells[testCellKey];
     });
   } // Pull user segments from options
 
@@ -50,16 +56,14 @@ function createSearchUrl(query, parameters, options) {
   }
 
   if (parameters) {
-    const {
-      page,
-      resultsPerPage,
-      filters,
-      sortBy,
-      sortOrder,
-      section,
-      collectionId,
-      fmtOptions
-    } = parameters; // Pull page from parameters
+    var page = parameters.page,
+        resultsPerPage = parameters.resultsPerPage,
+        filters = parameters.filters,
+        sortBy = parameters.sortBy,
+        sortOrder = parameters.sortOrder,
+        section = parameters.section,
+        collectionId = parameters.collectionId,
+        fmtOptions = parameters.fmtOptions; // Pull page from parameters
 
     if (!helpers.isNil(page)) {
       queryParams.page = page;
@@ -103,10 +107,10 @@ function createSearchUrl(query, parameters, options) {
 
   queryParams._dt = Date.now();
   queryParams = helpers.cleanParams(queryParams);
-  const queryString = qs.stringify(queryParams, {
+  var queryString = qs.stringify(queryParams, {
     indices: false
   });
-  return `${serviceUrl}/search/${encodeURIComponent(query)}?${queryString}`;
+  return "".concat(serviceUrl, "/search/").concat(encodeURIComponent(query), "?").concat(queryString);
 }
 /**
  * Interface to search related API calls
@@ -117,8 +121,10 @@ function createSearchUrl(query, parameters, options) {
  */
 
 
-class Search {
-  constructor(options) {
+var Search = /*#__PURE__*/function () {
+  function Search(options) {
+    _classCallCheck(this, Search);
+
     this.options = options || {};
     this.eventDispatcher = new EventDispatcher(options.eventDispatcher);
   }
@@ -139,48 +145,56 @@ class Search {
    */
 
 
-  getSearchResults(query, parameters) {
-    let requestUrl;
-    const fetch = this.options && this.options.fetch || fetchPonyfill({
-      Promise
-    }).fetch;
+  _createClass(Search, [{
+    key: "getSearchResults",
+    value: function getSearchResults(query, parameters) {
+      var _this = this;
 
-    try {
-      requestUrl = createSearchUrl(query, parameters, this.options);
-    } catch (e) {
-      return Promise.reject(e);
-    }
+      var requestUrl;
+      var fetch = this.options && this.options.fetch || fetchPonyfill({
+        Promise: Promise
+      }).fetch;
 
-    return fetch(requestUrl).then(response => {
-      if (response.ok) {
-        return response.json();
+      try {
+        requestUrl = createSearchUrl(query, parameters, this.options);
+      } catch (e) {
+        return Promise.reject(e);
       }
 
-      return helpers.throwHttpErrorFromResponse(new Error(), response);
-    }).then(json => {
-      // Search results
-      if (json.response && json.response.results) {
-        if (json.result_id) {
-          json.response.results.forEach(result => {
-            // eslint-disable-next-line no-param-reassign
-            result.result_id = json.result_id;
-          });
+      return fetch(requestUrl).then(function (response) {
+        if (response.ok) {
+          return response.json();
         }
 
-        this.eventDispatcher.queue('search.getSearchResults.completed', json);
-        return json;
-      } // Redirect rules
+        return helpers.throwHttpErrorFromResponse(new Error(), response);
+      }).then(function (json) {
+        // Search results
+        if (json.response && json.response.results) {
+          if (json.result_id) {
+            json.response.results.forEach(function (result) {
+              // eslint-disable-next-line no-param-reassign
+              result.result_id = json.result_id;
+            });
+          }
+
+          _this.eventDispatcher.queue('search.getSearchResults.completed', json);
+
+          return json;
+        } // Redirect rules
 
 
-      if (json.response && json.response.redirect) {
-        this.eventDispatcher.queue('search.getSearchResults.completed', json);
-        return json;
-      }
+        if (json.response && json.response.redirect) {
+          _this.eventDispatcher.queue('search.getSearchResults.completed', json);
 
-      throw new Error('getSearchResults response data is malformed');
-    });
-  }
+          return json;
+        }
 
-}
+        throw new Error('getSearchResults response data is malformed');
+      });
+    }
+  }]);
+
+  return Search;
+}();
 
 module.exports = Search;
