+++ lib/utils/helpers.js
@@ -1,20 +1,22 @@
-/* eslint-disable no-param-reassign */
-const qs = require('qs');
+"use strict";
 
-const utils = {
-  ourEncodeURIComponent: str => {
+/* eslint-disable no-param-reassign */
+var qs = require('qs');
+
+var utils = {
+  ourEncodeURIComponent: function ourEncodeURIComponent(str) {
     if (str) {
-      const parsedStrObj = qs.parse(`s=${str.replace(/&/g, '%26')}`);
+      var parsedStrObj = qs.parse("s=".concat(str.replace(/&/g, '%26')));
       parsedStrObj.s = parsedStrObj.s.replace(/\s/g, ' ');
       return qs.stringify(parsedStrObj).split('=')[1];
     }
 
     return null;
   },
-  cleanParams: paramsObj => {
-    const cleanedParams = {};
-    Object.keys(paramsObj).forEach(paramKey => {
-      const paramValue = paramsObj[paramKey];
+  cleanParams: function cleanParams(paramsObj) {
+    var cleanedParams = {};
+    Object.keys(paramsObj).forEach(function (paramKey) {
+      var paramValue = paramsObj[paramKey];
 
       if (typeof paramValue === 'string') {
         // Replace non-breaking spaces (or any other type of spaces caught by the regex)
@@ -26,14 +28,18 @@ const utils = {
     });
     return cleanedParams;
   },
-  throwHttpErrorFromResponse: (error, response) => response.json().then(json => {
-    error.message = json.message;
-    error.status = response.status;
-    error.statusText = response.statusText;
-    error.url = response.url;
-    error.headers = response.headers;
-    throw error;
-  }),
-  isNil: value => value == null
+  throwHttpErrorFromResponse: function throwHttpErrorFromResponse(error, response) {
+    return response.json().then(function (json) {
+      error.message = json.message;
+      error.status = response.status;
+      error.statusText = response.statusText;
+      error.url = response.url;
+      error.headers = response.headers;
+      throw error;
+    });
+  },
+  isNil: function isNil(value) {
+    return value == null;
+  }
 };
 module.exports = utils;
