+++ src/modules/autocomplete.js
@@ -2,7 +2,6 @@
 const qs = require('qs');
 const fetchPonyfill = require('fetch-ponyfill');
 const Promise = require('es6-promise');
-const EventDispatcher = require('../utils/event-dispatcher');
 const { throwHttpErrorFromResponse, cleanParams } = require('../utils/helpers');
 
 // Create URL from supplied query (term) and parameters
@@ -84,7 +83,6 @@ function createAutocompleteUrl(query, parameters, options) {
 class Autocomplete {
   constructor(options) {
     this.options = options || {};
-    this.eventDispatcher = new EventDispatcher(options.eventDispatcher);
   }
 
   /**
@@ -133,8 +131,6 @@ class Autocomplete {
               }
             });
           }
-
-          this.eventDispatcher.queue('autocomplete.getAutocompleteResults.completed', json);
 
           return json;
         }
