+++ src/modules/recommendations.js
@@ -2,7 +2,6 @@
 const qs = require('qs');
 const fetchPonyfill = require('fetch-ponyfill');
 const Promise = require('es6-promise');
-const EventDispatcher = require('../utils/event-dispatcher');
 const helpers = require('../utils/helpers');
 
 // Create URL from supplied parameters
@@ -75,7 +74,6 @@ function createRecommendationsUrl(podId, parameters, options) {
 class Recommendations {
   constructor(options) {
     this.options = options || {};
-    this.eventDispatcher = new EventDispatcher(options.eventDispatcher);
   }
 
   /**
@@ -120,8 +118,6 @@ class Recommendations {
               result.result_id = json.result_id;
             });
           }
-
-          this.eventDispatcher.queue('recommendations.getRecommendations.completed', json);
 
           return json;
         }
