+++ test/modules/browse.js
@@ -1,103 +1,87 @@
-"use strict";
-
-function _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError("Cannot call a class as a function"); } }
-
-function _defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if ("value" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } }
-
-function _createClass(Constructor, protoProps, staticProps) { if (protoProps) _defineProperties(Constructor.prototype, protoProps); if (staticProps) _defineProperties(Constructor, staticProps); return Constructor; }
-
 /* eslint-disable object-curly-newline, no-underscore-dangle */
-var qs = require('qs');
+const qs = require('qs');
+const fetchPonyfill = require('fetch-ponyfill');
+const Promise = require('es6-promise');
+const helpers = require('../utils/helpers');
 
-var fetchPonyfill = require('fetch-ponyfill');
-
-var Promise = require('es6-promise');
-
-var EventDispatcher = require('../utils/event-dispatcher');
-
-var helpers = require('../utils/helpers'); // Create URL from supplied filter name, value and parameters
-
-
+// Create URL from supplied filter name, value and parameters
 function createBrowseUrl(filterName, filterValue, parameters, options) {
-  var apiKey = options.apiKey,
-      version = options.version,
-      serviceUrl = options.serviceUrl,
-      sessionId = options.sessionId,
-      clientId = options.clientId,
-      userId = options.userId,
-      segments = options.segments,
-      testCells = options.testCells;
-  var queryParams = {
-    c: version
-  };
+  const {
+    apiKey,
+    version,
+    serviceUrl,
+    sessionId,
+    clientId,
+    userId,
+    segments,
+    testCells,
+  } = options;
+  let queryParams = { c: version };
+
   queryParams.key = apiKey;
   queryParams.i = clientId;
-  queryParams.s = sessionId; // Validate filter name is provided
+  queryParams.s = sessionId;
 
+  // Validate filter name is provided
   if (!filterName || typeof filterName !== 'string') {
     throw new Error('filterName is a required parameter of type string');
-  } // Validate filter value is provided
+  }
 
-
+  // Validate filter value is provided
   if (!filterValue || typeof filterValue !== 'string') {
     throw new Error('filterValue is a required parameter of type string');
-  } // Pull test cells from options
+  }
 
-
+  // Pull test cells from options
   if (testCells) {
-    Object.keys(testCells).forEach(function (testCellKey) {
-      queryParams["ef-".concat(testCellKey)] = testCells[testCellKey];
+    Object.keys(testCells).forEach((testCellKey) => {
+      queryParams[`ef-${testCellKey}`] = testCells[testCellKey];
     });
-  } // Pull user segments from options
+  }
 
-
+  // Pull user segments from options
   if (segments && segments.length) {
     queryParams.us = segments;
-  } // Pull user id from options
+  }
 
-
+  // Pull user id from options
   if (userId) {
     queryParams.ui = userId;
   }
 
   if (parameters) {
-    var page = parameters.page,
-        resultsPerPage = parameters.resultsPerPage,
-        filters = parameters.filters,
-        sortBy = parameters.sortBy,
-        sortOrder = parameters.sortOrder,
-        section = parameters.section,
-        fmtOptions = parameters.fmtOptions; // Pull page from parameters
+    const { page, resultsPerPage, filters, sortBy, sortOrder, section, fmtOptions } = parameters;
 
+    // Pull page from parameters
     if (!helpers.isNil(page)) {
       queryParams.page = page;
-    } // Pull results per page from parameters
+    }
 
-
+    // Pull results per page from parameters
     if (!helpers.isNil(resultsPerPage)) {
       queryParams.num_results_per_page = resultsPerPage;
     }
 
     if (filters) {
       queryParams.filters = filters;
-    } // Pull sort by from parameters
+    }
 
-
+    // Pull sort by from parameters
     if (sortBy) {
       queryParams.sort_by = sortBy;
-    } // Pull sort order from parameters
+    }
 
-
+    // Pull sort order from parameters
     if (sortOrder) {
       queryParams.sort_order = sortOrder;
-    } // Pull section from parameters
+    }
 
-
+    // Pull section from parameters
     if (section) {
       queryParams.section = section;
-    } // Pull format options from parameters
+    }
 
-
+    // Pull format options from parameters
     if (fmtOptions) {
       queryParams.fmt_options = fmtOptions;
     }
@@ -105,11 +89,12 @@ function createBrowseUrl(filterName, filterValue, parameters, options) {
 
   queryParams._dt = Date.now();
   queryParams = helpers.cleanParams(queryParams);
-  var queryString = qs.stringify(queryParams, {
-    indices: false
-  });
-  return "".concat(serviceUrl, "/browse/").concat(encodeURIComponent(filterName), "/").concat(encodeURIComponent(filterValue), "?").concat(queryString);
+
+  const queryString = qs.stringify(queryParams, { indices: false });
+
+  return `${serviceUrl}/browse/${encodeURIComponent(filterName)}/${encodeURIComponent(filterValue)}?${queryString}`;
 }
+
 /**
  * Interface to browse related API calls
  *
@@ -117,15 +102,11 @@ function createBrowseUrl(filterName, filterValue, parameters, options) {
  * @inner
  * @returns {object}
  */
-
-
-var Browse = /*#__PURE__*/function () {
-  function Browse(options) {
-    _classCallCheck(this, Browse);
-
+class Browse {
+  constructor(options) {
     this.options = options || {};
-    this.eventDispatcher = new EventDispatcher(options.eventDispatcher);
   }
+
   /**
    * Retrieve browse results from API
    *
@@ -142,51 +123,40 @@ var Browse = /*#__PURE__*/function () {
    * @returns {Promise}
    * @see https://docs.constructor.io
    */
+  getBrowseResults(filterName, filterValue, parameters) {
+    let requestUrl;
+    const fetch = (this.options && this.options.fetch) || fetchPonyfill({ Promise }).fetch;
 
+    try {
+      requestUrl = createBrowseUrl(filterName, filterValue, parameters, this.options);
+    } catch (e) {
+      return Promise.reject(e);
+    }
 
-  _createClass(Browse, [{
-    key: "getBrowseResults",
-    value: function getBrowseResults(filterName, filterValue, parameters) {
-      var _this = this;
-
-      var requestUrl;
-      var fetch = this.options && this.options.fetch || fetchPonyfill({
-        Promise: Promise
-      }).fetch;
-
-      try {
-        requestUrl = createBrowseUrl(filterName, filterValue, parameters, this.options);
-      } catch (e) {
-        return Promise.reject(e);
-      }
-
-      return fetch(requestUrl).then(function (response) {
+    return fetch(requestUrl)
+      .then((response) => {
         if (response.ok) {
           return response.json();
         }
 
         return helpers.throwHttpErrorFromResponse(new Error(), response);
-      }).then(function (json) {
+      })
+      .then((json) => {
         if (json.response && json.response.results) {
           if (json.result_id) {
             // Append `result_id` to each result item
-            json.response.results.forEach(function (result) {
+            json.response.results.forEach((result) => {
               // eslint-disable-next-line no-param-reassign
               result.result_id = json.result_id;
             });
           }
-
-          _this.eventDispatcher.queue('browse.getBrowseResults.completed', json);
 
           return json;
         }
 
         throw new Error('getBrowseResults response data is malformed');
       });
-    }
-  }]);
-
-  return Browse;
-}();
+  }
+}
 
 module.exports = Browse;
