+++ test/utils/event-dispatcher.js
@@ -1,20 +1,8 @@
-"use strict";
-
-function _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError("Cannot call a class as a function"); } }
-
-function _defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if ("value" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } }
-
-function _createClass(Constructor, protoProps, staticProps) { if (protoProps) _defineProperties(Constructor.prototype, protoProps); if (staticProps) _defineProperties(Constructor, staticProps); return Constructor; }
-
 /* eslint-disable no-unneeded-ternary */
-var helpers = require('../utils/helpers');
+const helpers = require('../utils/helpers');
 
-var EventDispatcher = /*#__PURE__*/function () {
-  function EventDispatcher(options) {
-    var _this = this;
-
-    _classCallCheck(this, EventDispatcher);
-
+class EventDispatcher {
+  constructor(options) {
     this.events = [];
     this.enabled = options && options.enabled === false ? false : true; // Defaults to 'true'
 
@@ -39,44 +27,40 @@ var EventDispatcher = /*#__PURE__*/function () {
       // - Important for the case where client library instantiated before beacon has loaded
 
 
-      helpers.addEventListener('cio.beacon.loaded', function () {
-        if (_this.enabled) {
-          _this.active = true;
-
-          _this.dispatchEvents();
+      helpers.addEventListener('cio.beacon.loaded', () => {
+        if (this.enabled) {
+          this.active = true;
+          this.dispatchEvents();
         }
       });
     }
   } // Push event data to queue
 
 
-  _createClass(EventDispatcher, [{
-    key: "queue",
-    value: function queue(name, data) {
-      this.events.push({
-        name: name,
-        data: data
-      });
+  queue(name, data) {
+    this.events.push({
+      name,
+      data
+    });
 
-      if (this.active) {
-        this.dispatchEvents();
-      }
-    } // Dispatch all custom events within queue on `window` of supplied name with data
-
-  }, {
-    key: "dispatchEvents",
-    value: function dispatchEvents() {
-      while (this.events.length) {
-        var item = this.events.shift();
-        var name = item.name,
-            data = item.data;
-        var eventName = "cio.client.".concat(name);
-        helpers.dispatchEvent(helpers.createCustomEvent(eventName, data));
-      }
+    if (this.active) {
+      this.dispatchEvents();
     }
-  }]);
+  } // Dispatch all custom events within queue on `window` of supplied name with data
 
-  return EventDispatcher;
-}();
+
+  dispatchEvents() {
+    while (this.events.length) {
+      const item = this.events.shift();
+      const {
+        name,
+        data
+      } = item;
+      const eventName = `cio.client.${name}`;
+      helpers.dispatchEvent(helpers.createCustomEvent(eventName, data));
+    }
+  }
+
+}
 
 module.exports = EventDispatcher;
