+++ test/utils/store.overflow.js
@@ -1,5 +1,3 @@
-"use strict";
-
 /* eslint-disable */
 
 /**
